+++ code/porting/profile-kernel.cu
@@ -19,7 +19,7 @@ int main(int argc, char *argv[]) {
     size_t numIterations = 8;
 
     size_t *data;
-    checkCudaError(cudaMallocManaged(&data, numElements * sizeof(double)));
+    checkCudaError(cudaMallocManaged(&data, numElements * sizeof(size_t)));
 
     initializeData(data, numElements);
