+++ code/porting/sol-prefetch.cu
@@ -19,15 +19,15 @@ int main(int argc, char *argv[]) {
     size_t numIterations = 8;
 
     size_t *data;
-    checkCudaError(cudaMallocManaged(&data, numElements * sizeof(double)));
+    checkCudaError(cudaMallocManaged(&data, numElements * sizeof(size_t)));
 
     initializeData(data, numElements);
 
     int device;
     cudaGetDevice(&device);
 
-    //# old version: checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(double), device));
-    checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(double), { .type = cudaMemLocationTypeDevice, .id = device }, 0));
+    //# old version: checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(size_t), device));
+    checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(size_t), { .type = cudaMemLocationTypeDevice, .id = device }, 0));
 
     //# main 'work'
     for (int it = 0; it < numIterations; ++it) {
@@ -37,8 +37,8 @@ int main(int argc, char *argv[]) {
     }
     checkCudaError(cudaDeviceSynchronize(), true);
 
-    //# old version: checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(double), cudaCpuDeviceId));
-    checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(double), { .type = cudaMemLocationTypeHost }, 0));
+    //# old version: checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(size_t), cudaCpuDeviceId));
+    checkCudaError(cudaMemPrefetchAsync(data, numElements * sizeof(size_t), { .type = cudaMemLocationTypeHost }, 0));
 
     verifyData(data, numElements, numIterations);
